+++ backend/graphql/DefaultState.js
@@ -7,25 +7,31 @@ import {
 import Matrix from "./Matrix";
 import DefaultStateVideoConnection from "./DefaultStateVideoConnection";
 import DefaultStateKwmConnection from "./DefualtStateKwmConnection";
+import { valueOf, valuesOf } from "./immutable";
 
 
 export default new GraphQLObjectType({
 	name: "DefaultState",
 	fields: () => ({
 		id: {
-			type: GraphQLString
+			type: GraphQLString,
+			resolve: state => String(valueOf(state, "id"))
 		},
 		slug: {
-			type: GraphQLString
+			type: GraphQLString,
+			resolve: state => valueOf(state, "slug")
 		},
 		matrix: {
-			type: Matrix
+			type: Matrix,
+			resolve: state => valueOf(state, "matrix")
 		},
 		videoConnections: {
-			type: new GraphQLList(DefaultStateVideoConnection)
+			type: new GraphQLList(DefaultStateVideoConnection),
+			resolve: state => valuesOf(valueOf(state, "videoConnections"))
 		},
 		kwmConnections: {
-			type: new GraphQLList(DefaultStateKwmConnection)
+			type: new GraphQLList(DefaultStateKwmConnection),
+			resolve: state => valuesOf(valueOf(state, "kwmConnections"))
 		}
 	})
 });