+++ backend/graphql/DiagramScreenCpuPort.js
@@ -5,18 +5,22 @@ import {
 
 import DiagramScreen from "./DiagramScreen";
 import CpuPort from "./CpuPort";
+import { valueOf } from "./immutable";
 
 export default new GraphQLObjectType({
 	name: "DiagramScreenCpuPort",
 	fields: () => ({
 		id: {
-			type: GraphQLString
+			type: GraphQLString,
+			resolve: connection => String(valueOf(connection, "id"))
 		},
 		diagramScreen: {
 			type: DiagramScreen,
+			resolve: connection => valueOf(connection, "diagramScreen")
 		},
 		cpuPort: {
-			type: CpuPort
+			type: CpuPort,
+			resolve: connection => valueOf(connection, "cpuPort")
 		}
 	})
 });