+++ backend/graphql/WeeklyTimerQueries.js
@@ -9,13 +9,14 @@ import WeeklyTimer from "./WeeklyTimer";
 import {
 	db
 } from "../TaistoService";
+import { valuesOf } from "./immutable";
 
 export default {
 	weeklyTimers: {
 		name: "WeeklyTimers",
 		type: new GraphQLList(WeeklyTimer),
 		resolve: (_, args) => new Promise((resolve, reject) => {
-			resolve(db.weeklyTimers);
+			resolve(valuesOf(db.weeklyTimers));
 		})
 	},
 	weeklyTimerById: {
@@ -39,7 +40,7 @@ export default {
 			}
 		},
 		resolve: (_, args) => new Promise((resolve, reject) => {
-			resolve(db.weeklyTimers.find(p => p.slug === args.slug));
+			resolve(db.weeklyTimers.find(p => p.get("slug") === args.slug));
 		})
 	}
 }